refactor: 不要なインターフェース定義を削除しアーキテクチャを簡素化
- WorkflowRegistryInterface、WorkflowResolverInterfaceを削除
  - 同一パッケージ内で実装されているクラスにインターフェースは不要
- WorkflowQueueItem型をserverパッケージに移動
  - serverパッケージ固有の実装のため、coreから分離
- 各クラスを直接使用するように修正
